+++ src/com/donny/dendrofinance/gui/menu/util/acc/AppDepGui.java
@@ -309,7 +309,7 @@ public class AppDepGui extends RegisterFrame {
                         new LAccountSet("D!" + Account.depreciationName + "(" + x.abs() + "), " + accs, CURRENT_INSTANCE)
                 );
             }
-            CURRENT_INSTANCE.DATA_HANDLER.DATABASE.TRANSACTIONS.add(capsule, ImportHandler.ImportMode.KEEP);
+            CURRENT_INSTANCE.DATA_HANDLER.TRANSACTIONS.add(capsule, ImportHandler.ImportMode.KEEP);
             CALLER.updateTable();
             dispose();
         } catch (ValidationFailedException e) {