+++ src/com/donny/dendrofinance/gui/menu/util/acc/TaxZeroGui.java
@@ -89,7 +89,7 @@ public class TaxZeroGui extends RegisterFrame {
                     "Tax Zeroing",
                     new LAccountSet(sb.substring(2), CURRENT_INSTANCE)
             );
-            CURRENT_INSTANCE.DATA_HANDLER.DATABASE.TRANSACTIONS.add(capsule, ImportHandler.ImportMode.KEEP);
+            CURRENT_INSTANCE.DATA_HANDLER.TRANSACTIONS.add(capsule, ImportHandler.ImportMode.KEEP);
             CALLER.updateTable();
             dispose();
         } catch (ValidationFailedException e) {